pyproject.toml
[project]
name = "autovi-evaluation"
version = "3.0"
description = "Evaluation scripts for the AutoVI dataset"
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.10,<3.14"
dependencies = [
    "numpy",
    "pillow",
    "tqdm",
    "tabulate",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "faiss-cpu>=1.7.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "loguru>=0.7.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
